python_docker/requirements.txt
python-dotenv==0.15.0
redis==3.5.3